Catch :task_has_failed on the cli. (Fixes #88)
Since the server throws the Guard symbol :task_has_failed
to exit quickly, the cli wasn't catching it at all, leading to an
abrupt process termination and leaving the server running.

Now that :task_has_failed is catched, the cli runner tries to
run the specs anyway and terminates the server properly if
the runner is not available.