Output the runner response body when unavailable.
When checking if the Jasmine spec server is running,
we previously did just a HEAD request to get the status
code. To help debugging cases where the server fails,
the test request is now a normal GET request and we
can now output the response body for further
diagnosis. (See #98)