Bug fix: coverage hangs for large files
The code that used IO.pipe & ChildProcess to run istanbul was waiting for the process to finish before reading the result.  With a file whose instrumentation is large enough to exceed the system IO buffer size, the istanbul process would block and we'd have deadlock.  Rather than fixing that using the appropriate nonblocking reading, just switched to %x[]