Faster Jasmine runner exception detection. (See #112)
This adds a check to see if  Jasmine is available by looking
at `window.jasmine`. If so, the normal waiting for the spec
results happens until the given timeout is reached. If Jasmine
is missing, then it's very likely that an exception is show by
the Rails error page (which of course doesn't include Jasmine).
In that case fail the specs with the body text as error message,
so that diagnosis directly in the console is possible.