Fix detect a directory as the phantomjs executable
`File.executable?(exe)` will return `true` for directories. This patch ensures that only files are selected.

With the existing code, it can detect a directory named `phantomjs` as the phantomjs executable. This was exactly what was happening in my case: Sencha Cmd tool creates a "phantomjs" directory which is reachable from the path, and this was making guard-jasmine fail.