New reporter that supports Jasmine 2 API
This adds a new reporter that uses the Jasmine 2 JsApiReporter API.

It creates an almost identical JSON response to the
old API but with a few differences.

The main one being that it's unable to create the "messages" array for each spec.

Instead it puts the messages on the "errors" array,
but has more detail about the expectation that failed.

It's a first pass at resolving # 161