Runner as a class with options as an instance var
This converts the Runner module to a class with it's options stored
as an instance variable, which can be over-ridden on a per-run basis.

This is intended to simplify the runner code by removing the need to pass
the options to every single method.  It's also a first step towards
improving the logic of when a given message is hidden/shown in the runner.